Guillaume, les test sont toujours KO je pense, sinon fix struct padding

+++ srcs/exp/machine_expansion.c
@@ -25,7 +25,7 @@ u_int8_t		check_tilde_path_exp
 	tmp[0] = ft_strsub(str, 0, i);
 	i = (str[i] == '~' && str[i + 1]
 		&& (str[i + 1] == '-' || str[i + 1] == '+')) ? i + 2 : i + 1;
-	tmp[1] = ft_strsub(str, i, ft_strlen(str) - i);
+	tmp[1] = ft_strsub(str, (unsigned int)i, ft_strlen(str) - i);
 	if ((tmp[1][0] == '/' || tmp[1][0] == ':') && !*tmp[0])
 	{
 		ft_strdel(&tmp[0]);
